requirements.txt
numpy<2

streamlit
openai
azure-identity
requests
python-dotenv
pandas
matplotlib
pyodbc
sqlalchemy
sentence-transformers 
faiss-cpu 
datasets 
accelerate>=0.27.2

# Deepseek Coder fine-tuning dependencies
transformers==4.38.2
peft==0.8.2
datasets==2.18.0
bitsandbytes==0.42.0
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
pyjwt
fastembed
qdrant-client 

azure-identity>=1.15.0
azure-storage-blob>=12.19.0
seaborn>=0.12.0
plotly>=5.17.0
PyYAML>=6.0